refactor(discovery): one file per discovery surface
Pure structural move (no behavior change): parameter_analysis/discovery.rs
(1590 lines) is now parameter_analysis/discovery/ —
- mod.rs      module docs + imports + check_discovery orchestrator + shared
              helpers (dedupe_reflection_params, explicit_param_names,
              merge/intersect_char_sets, json regexes)
- fragment.rs / query.rs / header.rs / path.rs / cookie.rs / form.rs
              one independent discovery surface each

Surfaces inherit the shared helpers + imports via `use super::*`. Callers
keep using `discovery::check_discovery`, `discovery::dedupe_reflection_params`,
`discovery::explicit_param_names` unchanged. tests.rs untouched (only
COMMON_PROBE_HEADERS needed pub(super) so the test module still sees it).
28 discovery tests pass, clippy clean.
